fix(ui): single bottom playbar with live-updating progress and timer
- Remove inline replay controls from recordings panel (keep highlight
  and Playing/Paused label only)
- Add client-side interpolated progress to the bottom ReplayBar: uses
  server currentTime as anchor, ticks every 1s locally so both the
  progress bar and timer update smoothly between server events
- Use linear 1s CSS transition on progress fill for smooth animation
- Strip "recordings/" prefix from displayed filename
