+++ apps/ui/src/Controls/RecordReplay.module.css
@@ -175,81 +175,6 @@
   font-weight: 500;
 }
 
-/* ── Inline replay controls ── */
-
-.replayControls {
-  display: flex;
-  gap: var(--space-1);
-  flex-shrink: 0;
-}
-
-.transportBtn {
-  display: flex;
-  align-items: center;
-  justify-content: center;
-  width: 26px;
-  height: 26px;
-  border: none;
-  background: var(--surface-bg);
-  border-radius: var(--radius-sm);
-  cursor: pointer;
-  flex-shrink: 0;
-  transition: background var(--transition-fast);
-}
-
-.transportBtn:hover {
-  background: var(--surface-bg-hover);
-}
-
-.stopBtn {
-  composes: transportBtn;
-}
-
-.stopBtn:hover {
-  background: rgba(255, 68, 68, 0.12);
-}
-
-.transportIcon {
-  width: 12px;
-  height: 12px;
-  fill: var(--color-gray-light);
-}
-
-.stopBtn:hover .transportIcon {
-  fill: #f44;
-}
-
-/* ── Progress bar ── */
-
-.progressRow {
-  display: flex;
-  align-items: center;
-  gap: var(--space-3);
-  width: 100%;
-}
-
-.progressBar {
-  flex: 1;
-  height: 4px;
-  background: rgba(255, 255, 255, 0.08);
-  border-radius: 999px;
-  overflow: hidden;
-}
-
-.progressFill {
-  height: 100%;
-  background: var(--color-accent, rgba(59, 130, 246, 0.7));
-  border-radius: 999px;
-  transition: width 1s linear;
-}
-
-.progressTime {
-  font-size: var(--text-xs);
-  font-variant-numeric: tabular-nums;
-  color: var(--color-gray-light);
-  white-space: nowrap;
-  flex-shrink: 0;
-}
 
 .empty {
   font-size: var(--text-sm);
